SymaX Headless Mode
From: https://www.rcgroups.com/forums/showpost.php?p=38630343&postcount=10288

The wind died enough for me to investigate and test the flight performance of the SymaX code.

One of the issues is that while the actual control information coming from channels 1-4 is treated normally by the flight controller, the trim information which gets added to it is integrated over a long, roughly three second time. This makes for an interesting experience once one knows what's happening. Applying full ailerons takes those three seconds. At its maximum deflection and with full power, the quad will dive at about 75 degrees at a high rate of speed. Upon releasing the controls, it will continue the dive at a much less severe angle until the added trim is integrated away. The speeds which can be reached are much greater than usual if the angle and power are carefully managed.

The yaw does not seem to be affected by this.

This is pretty dramatic and fun but it's also quite different from the usual flight control experience, and so I have set up channel 6 (once used by SymaX to control rates) to turn on and off the addition of trim to the primary control channels, using a flag labeled SYMAX_XTRM_RATES.

A complication arises when headless mode is used. If the headless direction is away from the pilot and the quad is facing to his right, right stick will result in the quad moving to the right. However, as the trim kicks in it will still be relative to the quad and it will begin pushing the quad towards the pilot. Upon release of the sticks, the quad's movement to the pilot's right will stop but the movement towards the pilot will continue until the trim dissipates. The path of the quad will be a broad L.

While the quad is pointed in the headless direction, the trim will act in concert with the applied controls, but if it is facing opposite it will fight the controls and result in very reduced control rates.

There are other oddities, one of which arises when holding the quad steady in a breeze where it is not facing the headless direction. The stick controls to hold it in place will cause the integration of trim in the flight controller such that one will have to hold quite a lot of deflection to keep it in place, and if one releases the controls the quad will bank 30 degrees in the downwind direction and fly off!

Further complicating the use of headless mode with the added trims is the fact that the severe angles encountered while flying this way cause a rapid drift of the gyro responsible for headless mode, so that in a short time the controls are relative to a new headless direction! Even an experienced pilot who has figured all this out will have trouble calculating how to put all this together in such a way as to bring the quad to the desired position.

For this reason I am suggesting the following patch which allows the addition of trim to be turned on and off via channel 6, and which turns off the addition of trim when headless mode is active. One could still have fun turning one's Syma quads into very fast devices capable of extreme angles but with an oddly slow component of their control, but the default would be more traditional. And if one chooses to use headless mode, the complexity of figuring out what will happen when a particular set of controls are applied and added to a set of trims slowly increasing and decreasing in different and changing directions is eliminated.

I'd also like to say that while it is nice to be able to turn off the addition of trim to the primary controls, on the whole it's a great idea and makes the Syma quads much more fun and exciting. Thanks for creating this amazing and fun system.